pyproject.toml
[tool.poetry]
name = "django-filer-image"
version = "0.0.1"
description = ""
authors = ["Дмитрий <acrius@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.5"
django = "^1.11 || ^2.0"
pillow = "^6.1.0"
django-filer = "^1.5"
funcy = "^1.13"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
